pyproject.toml
[project]
name = "platonic-universe"
version = "0.1.0"
description = "Scripts to test the Platonic Representation Hypothesis on astronomical data"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "argparse>=1.4.0",
    "astropt>=2.0.5",
    "llvmlite>=0.44.0",
    "numpy>=2.3.2",
    "opencv-python>=4.11.0.86",
    "pillow>=11.3.0",
    "polars>=1.32.2",
    "scikit-image>=0.25.2",
    "scikit-learn>=1.7.1",
    "torch>=2.8.0",
    "transformers>=4.55.0",
]

[build-system]  
requires = ["hatchling", "hatch-requirements-txt", "hatch-vcs"]
build-backend = "hatchling.build"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E501"]

[tool.uv]
package = true

[tool.hatch.build.targets.wheel]
packages = ["src/pu"]
